# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones CRUD y administrativas en EduGo",
        "title": "EduGo API Administración",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/v1",
    "paths": {
...
        "/stats/global": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stats"
                ],
                "summary": "Estadísticas globales",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/GlobalStatsResponse"
                        }
                    }
                }
            }
        },
        "/subjects": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjects"
                ],
                "summary": "Crear materia",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CreateSubjectResponse"
                        }
                    }
                }
            }
        },
        "/subjects/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Actualiza información de una materia del catálogo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Subjects"
                ],
                "summary": "Actualizar materia (Post-MVP)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la materia",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos a actualizar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateSubjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SubjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Solo administradores",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Materia no encontrada",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/units": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Units"
                ],
                "summary": "Crear unidad académica",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CreateUnitResponse"
                        }
                    }
                }
            }
        },
        "/units/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Units"
                ],
                "summary": "Actualizar unidad",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la unidad",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    }
                }
            }
        },
        "/units/{id}/members": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Units"
                ],
                "summary": "Asignar membresía",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la unidad",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/SuccessResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateSubjectResponse": {
            "type": "object",
            "properties": {
                "subject_id": {
                    "type": "string",
                    "example": "subject-uuid-123"
                }
            }
        },
        "CreateUnitResponse": {
            "type": "object",
            "properties": {
                "unit_id": {
                    "type": "string",
                    "example": "unit-uuid-123"
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request"
                },
                "message": {
                    "type": "string",
                    "example": "Field 'guardian_id' is required"
                }
            }
        },
        "GlobalStatsResponse": {
            "type": "object",
            "properties": {
                "active_users_30d": {
                    "type": "integer",
                    "example": 980
                },
                "total_materials": {
                    "type": "integer",
                    "example": 450
                },
                "total_users": {
                    "type": "integer",
                    "example": 1250
                }
            }
        },
...
        "SubjectResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-01-01T00:00:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Curso de matemáticas"
                },
                "id": {
                    "type": "string",
                    "example": "uuid-subject-101"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "example": "Matemáticas"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-01-15T10:00:00Z"
                }
            }
        },
        "SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación exitosa"
                }
            }
        },
        "UpdateSubjectRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Curso avanzado de matemáticas"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "name": {
                    "type": "string",
                    "example": "Matemáticas Avanzadas"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}